feat: edit CA page modified
User verification is added to the edit CA page. If the user is not logged in, he/she will be redirected to the login page. If the user is logged in, but not the owner of the CA, he/she will be redirected to the CA list page.
AppWrap component removed from the edit CA page. Sidebar component added to the edit CA page.
Credit field enabled in the edit CA page.
Bug fixed in the edit CA page. The ref_code and college fields were of the same name. This caused the ref_code field to change its value to the college field's value. This bug is fixed.