feat: Sidebar updated
Sidebar is shown only when the window width is greater than 900px. Otherwise, it is hidden and a hamburger menu is shown. When the hamburger menu is clicked, the sidebar is shown.
Sidebar min-height is set to 100vh and height is set to 100%. This is done to make sure that the sidebar is always shown completely even when the content is less than the height of the window.